+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [v3.1.13](https://github.com/beliaev-maksim/linux_hpc_launcher_slurm/compare/v3.1.12...v3.1.13)
+* enable product improvement by default
+* remove unused function
+* added `--debug` flag for future expansions
+* print `scancel` to terminal on cancel
+
 ## [v3.1.12](https://github.com/beliaev-maksim/linux_hpc_launcher_slurm/compare/v3.1.11...v3.1.12)
 * prepend HOSTNAME to DISPLAY environment variable, 
 eg if `DISPLAY=:3`, it will be converted to `DISPLAY=ottvnc3.ansys.com:3`